backend/requirements.txt
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
Flask==3.1.3
flask-cors==6.0.2
gunicorn==25.3.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
numpy==2.4.4
packaging==26.0
pandas==3.0.1
PyJWT==2.12.1
pymongo==4.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.33.1
resend==2.26.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.7
